Backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
motor==3.6.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.9.2
pydantic-settings>=2.10.1
authlib==1.3.1
python-dotenv==1.0.1

